requirements.txt
streamlit
pandas
geopandas
folium
matplotlib
fitparse
shapely
python-dotenv
garth